+++ demo/SecsFrame.CommunicationDemo/Models/ActivityExportClassification.cs
@@ -1,0 +1,7 @@
+namespace SecsFrame.CommunicationDemo.Models;
+
+internal enum ActivityExportClassification
+{
+    MetadataOnly,
+    RedactedContent,
+}